requirements.txt
# Core
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Deep Learning
torch>=2.1.0
torchvision>=0.16.0
ultralytics>=8.0.0  # YOLOv8
onnxruntime-gpu>=1.16.0

# Computer Vision
opencv-python>=4.8.0
Pillow>=10.0.0
albumentations>=1.3.0

# Face Recognition
insightface>=0.7.0
facenet-pytorch>=2.5.0

# CLIP
transformers>=4.35.0
sentence-transformers>=2.2.0

# Vector Database
pymilvus>=2.3.0

# Database
sqlalchemy>=2.0.23
asyncpg>=0.29.0
redis>=5.0.0

# Image Processing
imageio>=2.31.0
scikit-image>=0.22.0

# Video Processing
av>=10.0.0
ffmpeg-python>=0.2.0

# Monitoring
prometheus-client>=0.19.0
structlog>=23.2.0

# Utils
numpy>=1.24.0
python-dotenv>=1.0.0
aiofiles>=23.2.0
python-multipart>=0.0.6
httpx>=0.25.0
